tech(engine): Rename Action pause/resume/stop as Command
