+++ engine/src/main/java/com/chutneytesting/engine/domain/execution/ExecutionManager.java
@@ -1,8 +1,8 @@
 package com.chutneytesting.engine.domain.execution;
 
-import com.chutneytesting.engine.domain.execution.action.PauseExecutionAction;
-import com.chutneytesting.engine.domain.execution.action.ResumeExecutionAction;
-import com.chutneytesting.engine.domain.execution.action.StopExecutionAction;
+import com.chutneytesting.engine.domain.execution.command.PauseExecutionCommand;
+import com.chutneytesting.engine.domain.execution.command.ResumeExecutionCommand;
+import com.chutneytesting.engine.domain.execution.command.StopExecutionCommand;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -15,16 +15,16 @@ public class ExecutionManager {
 
     public void pauseExecution(Long executionId) {
         LOGGER.info("Pause requested for " + executionId);
-        RxBus.getInstance().post(new PauseExecutionAction(executionId));
+        RxBus.getInstance().post(new PauseExecutionCommand(executionId));
     }
 
     public void resumeExecution(Long executionId) {
         LOGGER.info("Resume requested for " + executionId);
-        RxBus.getInstance().post(new ResumeExecutionAction(executionId));
+        RxBus.getInstance().post(new ResumeExecutionCommand(executionId));
     }
 
     public void stopExecution(Long executionId) {
         LOGGER.info("Stop requested for " + executionId);
-        RxBus.getInstance().post(new StopExecutionAction(executionId));
+        RxBus.getInstance().post(new StopExecutionCommand(executionId));
     }
 }
