+++ engine/src/main/java/com/chutneytesting/engine/domain/execution/command/ResumeExecutionCommand.java
@@ -1,12 +1,12 @@
-package com.chutneytesting.engine.domain.execution.action;
+package com.chutneytesting.engine.domain.execution.command;
 
 import com.chutneytesting.engine.domain.execution.event.Event;
 
-public class StopExecutionAction implements Event {
+public class ResumeExecutionCommand implements Event {
 
     private final Long executionId;
 
-    public StopExecutionAction(Long executionId) {
+    public ResumeExecutionCommand(Long executionId) {
         this.executionId = executionId;
     }
